Remove bit_tools and move tests into data_types
